+++ src/pages/main/productRenderer.ts
